backend/requirements.txt
# 环境所需依赖

django==2.0.6
django-cors-headers==3.2.0
django-filter==2.2.0
djangorestframework==3.10.0
djangorestframework-jwt 1.11.0

mysqlclient==2.0.3
django-redis==4.7.0

pillow==8.2.0
requests==2.25.1

https://codeload.github.com/sshwsfc/xadmin/zip/django2
